Update patch set 6
Patch Set 6:

> And why are we not fixing this in ShardDataTree.processNextTransaction(), where we already have a LOG.warn() and hence could enrich the text of that OptimisticLockFailedException?

My use case is investigation of a failure in SingleFeatureTest while building Distribution. The exception is visible near the end of job console output, the WARN is buried in the middle of big surefire report.

Do you have a use case where the user would NOT wish to have the path value already inserted into the DataValidationFailedException string?

Patch-set: 6
Label: Verified=0