BUG-8043: correct LengthConstraint definition
Each type has at most one length constraint attached to it, reflect
that fact in yang-model-api hiearchy.

This has implications on LengthStatement: it should not be exposing
LengthConstraints, as that is not what the argument holds. Define
ValueRange construct to hold the parsed-out (but unresolved) ranges
as parsed from length argument.

Change-Id: I74dd4a0951159d663ee19c7177e8cb2a43bf92d2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>