Make SimpleDateFormatUtil non-lenient
Converting an illegal revision, like 2017-16-05, results in incorrect
results in that the resulting Date will reflect 2018-04-05, which is
no a reversible transformation -- leading to hard to fix issues when
a String and Date representation meet.

Make sure we do not allow such invalid revisions through and fixup
bad test models.

Change-Id: I138cb5357cd6898bd64b745fbfed5d05b593e1e1
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>