Add the parent module of a submodule to the set of required lib sources
This issue is related to the following use case of YANG statement parser:
If we add a YANG submodule as a main source and its parent module
as a library source, YANG statement parser does not add the parent module
to the set of requires lib sources. This causes an Exception as the parser
cannot resolve belongs-to statement in the submodule.

This patch fixes the issue.

Change-Id: Icbfddbbbecec310f0f7de17456e6a12dab8167d6
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>