Avoid module namespace/submodule name conflict when using lib sources
The issue comes up when using yang-system-test artifact as a single
YANG file validator:

A parent module or a submodule is specified as its argument
and the same directory is specified as one of the library dirs
using the -p option.

The parent module/submodule ends up being added twice to the parsed
sources which causes a module namespace/submodule name collision later
in the parsing process.

Therefore remove the conflicting lib sources from the collection of
required lib sources before they cause any trouble.

Change-Id: Id41d9f57006adec7ff658904792f9bc319b8c11e
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>