Bug 7437: Fix yang-system-test logging without the --verbose option
yang-system-test artifact has the --verbose option which provides
more detailed logging output of the parsing process. However,
the logging ouput is full of DEBUG level messages coming from
the YANG statement parser even without the --verbose option.

This happens because the yang-system-test uses java.util.logging for
logging events and sets the level of logging only for that Logger.
However, the YANG statement parser uses ch.qos.logback backend for
logging, so we need to set the log level for that instead.

Therefore, replace the java.util.logging in yang-system-test with
ch.qos.logback backend.

Change-Id: I14ac775228506e4ed22cd1b55e7efec1898fa27f
Signed-off-by: Igor Foltin <igor.foltin@pantheon.tech>