Update patch set 1
Patch Set 1:

(2 comments)

> Maybe we should just deprecate these classes in yangtools and copy
 > them over to infrautils instead of copying between them and trying
 > to keep the best of both worlds in both. These were only put into
 > yangtools back when b/c it was the only place for common, reusable
 > code.

I'm personally all for that, but I'm afraid I don't think Robert is on board with that.. he does not want to have to make yangtools dependant on infrautils, right Robert?  Given that he's the yangtools commiter out of the 3 of us ;) discussing this here, I accept that problem as an axiom which we will not be able to resolve, or at least not on this Gerrit in the short term.  Therefore, I've proposed this change like that instead of what I understand would have been your preference Tom (given your -1 here).  (Now what we could and probably should do is instead then @deprecate and delegate to yangtools from infrautils for there; I'm happy to do that, if this gets merged.)

Robert, this request has some urgency in the context of the https://lists.opendaylight.org/pipermail/genius-dev/2017-September/001914.html discussion, where in genius/netvirt we are looking at using the Executors proposed here - but only if this goes in (we've already had too many problems with mysterious thread death not appearing in LOG and only "somewhere" [and it's never clear where, with systemd and ODL incontainers and what not in the picture]).  Therefore, I'd like to either get this accepted in yangtools, or I'll fork these classes (with this change) into infrautils to promote their usage, from infrautils instead of yangtools, in genius and netvirt.

Patch-set: 1
Label: Verified=0