Update patch set 2
Patch Set 2:

> I'm not saying make yangtools dependent on infrautils - @deprecate
 > and freeze the code here and copy functionality to infrautils.

Oh.  Misunderstood you there!  So.. wait, you ARE encouraging to fork this code - but why?  My starting point for any such matter has always been to try to avoid that as much as possible.. IMHO duplicated code is ALWAYS a Very Bad Idea.  Because things WILL, sooner or later, get fixed in one place, but not another, and things will re-diverge.. has that not been your experience also? So.. why? Are you anticipating that yangtools will eventually accept to depend on infrautils?  But then why make the fork now and not then?  Case in point with this change - I think it's BAD not good to have a different variety of SpecialExecutors in infrautils that diverged from the yangtools; e.g. because in infrautils it would have the LoggingThreadUncaughtExceptionHandler where the yangtools' one would not, if this does not get accepted; that's confusing, so why do that?  But if this does get merged, then.. there's really no urgency/need anymore to move these utilities in the short term?

Robert, thoughts about all this?

 > Now what we could and probably should do is instead then
 > @deprecate and delegate to yangtools from infrautils

Actually we cannot do this, I forget; we don't want to make infrautils dep on yangtools, either.

Patch-set: 2