Bug 9195 - Option to set supported deviations for SchemaContextFactory
There was missing API to set supported deviations in SchemaContextFactory.
This patch adds missing API and implementation for setting supported deviations
in SchemaContextFactory.
In order to avoid passing of numerous arguments into createSchemaContext(..) method
a concept of SchemaContextFactoryConfiguration has been introduced in this patch,
which encapsulates all options possible to set.

Change-Id: I8eef54d729b7bfed4cee41ee6758d969fc375dcf
Signed-off-by: Peter Kajsa <pkajsa@cisco.com>