BUG-4688: Eliminate use of ModuleIdentifier in ImpPrefixToSemVerModuleIdentifier
ModuleIdentifier being used to resolve semantic version is slightly
wrong, we already have SemVerSourceIdentifier, which serves exactly
the same purpose.

Rename ImpPrefixToSemVerModuleIdentifier namespace to better reflect
its use -- ImportPrefixToSemVerSourceIdentifier and adjust it to
carry SemVerSourceIdentifier.

This eliminates the need for ModuleIdentifier.getSemanticVersion()
and ModuleIdentifier.getNamespace() -- making it very obvious that
SourceIdentifier is the same thing with a different representation
of a revision.

Change-Id: Iac9948a271b1dcc09d8aeb3de4e019d3842a4562
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>