BUG-4688: Fix BelongsToPrefixToModuleIdentifier namespace
This namespace is used as a lookup indirection in the lookup of
QNameModule corresponding to a prefix:

String -> BelongsToPrefixToModuleIdentifier ->
    ModuleIdentifierToModuleQName -> QNameModule

Whereas we get the target module's StmtContext from the reactor,
and hence can go through ModuleCtxToModuleQName namespace instead:

String -> BelongsToPrefixToModuleCtx ->
    ModuleCtxToModuleQName -> QNameModule

This has the benefit of not needing any wildcard searches and
us getting precisely the QNameModule of the module which satisfied
our reactor requirement. It also removes the only source of
SimpleDateFormatUtil.DEFAULT_BELONGS_TO_DATE and hence that magic
value is removed, too.

Change-Id: I1d74bed41952ba7f15bfeb6af94193632b8fb8ab
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>