WIP: BUG-4688: Add IdentifierNamespaceKey
ModuleNamespace and SubmoduleNamespace require flexible lookups on
keys to arrive at a matching context, which is currently achieved
by magic constants and a special case in NamespaceSupport to make
it semi-work.

Add IdentifierNamespaceKey, which allows expressing loose matches
on keys of IdentifierNamespace. This allows us to properly express
lookup semantics and perform proper import resolution in reactor,
such that Prerequisites resolve to the real result instead of
re-searching the space for proper matches.

Change-Id: I900ed58404e7ab68a12d214ec7b1ebf11b60c103
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>