WIP: BUG-4688: Add flexible match support to ModelActionBuilder
This patch adds the support for flexible match of namespace keys
based on filter/reduce mechanics -- which are required to consistently
support revisionless and semantic imports without reliance on
magic constants in public interfaces.

In order to support required mechanics, we need to fix
the ValueAddedListener implementation and their lifecycle.

Change-Id: Icb22850e44de05d59820e1974e204dc1541e1f76
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>