BUG-4866: remove ImpPrefixToModuleIdentifier
ImpPrefixToModuleIdentifier is used only as indirection in
Prefix -> ModuleIdentifier -> QNameModule lookup. This is better
served via imported module context.

As it turns out, this namespace is very similar to ImportedModuleContext,
except it has a different key. Further investigation shows that
ImportedModuleContext users do not really care about the key, so
redefine ImportedModuleContext to be keyed by import prefix.

This eliminates the need for ModuleIdentifierToModuleQName namespace,
as there is nothing depending on its contents -- so remove it too,
lowering our footprint and simplifying code a bit.

Change-Id: I9804cd940fd732798cbd43194591637d021d1ac8
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>