BUG-4866: fix ImpPrefixToModuleIdentifier
ImpPrefixToModuleIdentifier is used only as indirection in
Prefix -> ModuleIdentifier -> QNameModule lookup. This is better
served via imported module context.

Rename the namespace to ImportPrefixToModuleCtx and have it return
imported StmtCtx directy, changing the lookup to work with
Prefix -> StmtCtx -> QNameModule.

This eliminates the need for ModuleIdentifierToModuleQName namespace,
as there is nothing depending on its contents -- so remove it too,
lowering our footprint and simplifying code a bit.

Change-Id: I9804cd940fd732798cbd43194591637d021d1ac8
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>