BUG-4688: eliminate SimpleDateFormatUtil.DEFAULT_DATE_REV
This is the final push for eliminating DEFAULT_DATE_REV, with all
other issues resolved, this removes the constant, adjusts the code
to correctly interpret non-present revisions and fixes up test
cases, eliminating the leaking 1970-01-01 string.

This actually fixes a few test cases, which relied on import with
revision 1970-01-01 matching modules which do not declare revision.

Change-Id: I59dbe8b06ad2d94e6a677959f29f5c2b52357d0b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>