YANGTOOLS-724: mark leaking fields
Lifecycle of 'augmenting' is wrong here, as we are resetting
this field after the effective statement has been built.

This indicates that the constructor is not getting enough/accurate
information about what is going on -- which is something we will
need to fix at some point.

Mark both occurances of this field, make the fields private and
expose a resetAugmenting() method.

Change-Id: Ica2678e885b624c850959fea21ee6120cfbd9da6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>