Add SchemaPath/SchemaNodeIdentifier.createChild(QName)
We can use a much shorter dispatch when we do not have to deal with
arrays/collections and have a single QName only. Introduce this
variant, which is widely used in YANG parser.

Introduce these methods and adjust some users of alternative methods
to take advantage of the new method.

Change-Id: If1d6288a6607b31bb45e9fd78c0319fb403e1528
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>