Merge PathUtils, TypeUtils and Utils methods
All of PathUtils and TypeUtils methods and some of Utils methods
are centered around dealing with statement arguments. Concentrate
all of them into a single utility class, rfc7950.stmt.ArgumentUtils.

This eliminates TypeUtils completely and leaves Utils with only
two methods, which are not used in statement reactor at all.

Change-Id: I816563e362a6294eed3df655be09d3d1d6173a04
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>