Update patch set 4
Patch Set 4:

releng team: we are in release-throttle for the SR relase. This patch is approved for merge as soon as the SR is out. This is not a regression, hence it should not cause an SR respin, I think.

If there is a respin, though, this fix should be included.

Patch-set: 4