Remove deprecated methods without loggerIdentity from SpecialExecutors
for https://jira.opendaylight.org/browse/CONTROLLER-1793

in order to force downstream projects (controller & mdsal) to use the
new methods when they upgrade to yangtools 2.0.0.  The thing is that we
cannot fix CONTROLLER-1793 in the current controller & mdsal before
switching to yangtools 2.0.0 anyway, because SpecialExecutors didn't
have the (my) previous change adding the loggerIdentity anyway.  So we
might as well do this in one round, instead of keeping these deprecated,
then once we are on yangtools 2.0.0 switch them, and then remove them?

Change-Id: I6a401979d71696536f23e1c5d1665149600f9876
Signed-off-by: Michael Vorburger <vorburger@redhat.com>