Update patch set 6
Patch Set 6:

Robert - the changes you made are fine however AbstractSchemaContext overrides the findModules(String) method and looks up in the namespaceToModules SetMultimap which doesn't preserve the revision ordering. So I removed the findModules methods from AbstractSchemaContext so the default implementation is used. The UTs in jsonrpc now pass.

Patch-set: 6