Patch request to a leaf list with JSON payload
A patch request to a leaf list with a JSON payload results in a ClasscastException.
More details about this bug at https://jira.opendaylight.org/browse/NETCONF-486
Removed the augmentationNode wrapper for mapEntryNodes as the ImutableMapNodeBuilder
expects a different heirarchy.
This made the JSON parsing to normalized node in parity with xml parsing.

Change-Id: I25d32acfe7797e0744b42ef7a5f46ab01971c2a8
Signed-off-by: Atul Gosain <agosain@luminanetworks.com>