Add CollectionWrappers
SchemaNode/SchemaContext has some unfortunate method contracts
which require returning a Set, while they are represented as
java.util.Map.values() in EffectiveStatement world, where they
support keyed lookups, which is really what users require.

To bridge these two worlds without duplicating the same information
in two collections we need a memory-efficient proxy to act as an
adaptor, as we know the primary storage (Map.values()) is known
to be both immutable and conforming to Set interface requirements.

This patch such a class.

Change-Id: I052e4f9d58aa1f054450840c2216b0fd2109bc61
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>