Add SchemaTreeAwareffectiveStatement
YANG defines two distinct ways to look at a YANG model:
- schema tree: The definition hierarchy specified within a module.
- data tree: An instantiated tree of any data modeled with YANG,
             e.g., configuration data, state data, combined configuration and
             state data, RPC or action input, RPC or action output, or
             notification.

This means that there really are two notions of 'a child node',
based on schema tree, or based on data tree. The data tree is
a strict subset of schema tree.

This patch introduces SchemaTreeAwareEffectiveStatement
and its companion SchemaTreeEffectiveStatement, updating extant
statements to support it.

JIRA: YANGTOOLS-853
Change-Id: Id2c41bb30c4376a846e4e974033b6d621ecfb155
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>