Update patch set 12
Patch Set 12:

> (1 comment)
Since case's substatements don't include rpcs nor notifications,  we should not take them account.
But  this change would be reverted as it's not correct:
we can not add another new namespace to hold case's children nodes, they should be put in the same one 'ChildSchemaNodeNamespace' as they are really at the same level with children of the closest ancestor, the collisions between which should be detected too.

Patch-set: 12
Label: Verified=0