WIP: Move search methods to NormalizedNodeContext
When we know we have a NormalizedNodeContext we can forgo bouncing
through the navigator to get a target node context. This simplifies
things a bit when we want to implement more complext logic.

This is not exactly extensible way of operation, but we know we will
only be evaluating on our context.

Change-Id: Ib8bea239f3193c54d3eacab903c75f9d43aee4e2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>