Fix SchemaContextUtil RevisionAwareXPath resolution
The mapping of yang-model-api SchemaNode hierarchy to YANG namespaces
is confusing, which is compounded by naming -- invoking notions
of equivalence when there is none.

This is very much evident when dealing with children identified by
QName in various contexts -- they usually work, but break down when
choice/case nodes are encountered, because their addressability is
different based on context -- they are addressable via 'node identifier',
but are invisble to XPath contexts.

Document the difference in extant DataNodeContainer and DataSchemaNode
API and introduce DataNodeContainer.findDataNode(), which follows
RFC7950 'data tree', not 'schema tree'.

Change-Id: Ic570da8dd97a686423c7e2b28e0db0d9c6907271
JIRA: YANGTOOLS-885
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>