Limit namespace lookps of import prefixes
Namespace definitions of ImpPrefixToNamespace, ImportPrefixToModuleCtx
and ImportPrefixToSemVerSourceIdentifier are wrong in that they include
search within other imported modules.

This is overreaching and results in false-positive lookups on these
namespaces -- i.e. even prefixes which are not declared in the local
source are lookup up successfully if a matching definition is found
in any of the modules which are imported by local module.

JIRA: YANGTOOLS-889
Change-Id: I28adefef4651a15b191915d7d50d5359d78cb9fe
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>