WIP: add dedicated XSD regex parser
We need to parse pattern strings using our own parser and translate
them to Pattern literals. This patch adds the grammar needed, as well
as the convertor.

FIXME: SpotBugs need to be suppressed
FIXME: && and CatEsc/ComplEsc need to be handled in charClass
FIXME: $ needs to be properly escaped
FIXME: XSDRegex needs to be plugged into RegexUtils
FIXME: validate Unicode block validity, expose a knob for strictness

JIRA: YANGTOOLS-887
Change-Id: I22e75ba2c4e7113ee742963851506eb7428e2b67
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>