Deprecate yang.binding.RpcImplementation
This is an old tagging interface, which is not used anywhere,
probably since we have started running binding RPCs on top
of DOM infrastructure.

Mark it deprecated for now, we will either remove it if it is not
useful or un-deprecate and documented it if we find a use for it.

Change-Id: If4f69c3202113d521a0992b1478247e2614ae523
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>