Fix LeafRefValidatation for refs defined in augmentations
LeafRefValidatation.validateNodeData traverses
children of a NormalizedNode
that contain leafrefs or are leafrefs' targets.

Information about referencing nodes and nodes being referenced
is stored in LeafRefContext under QName key.

But AugmentationNode has no QName, so validation was failing
with UnsupportedOperationException.

As a fix, validateNodeData is recursively invoked on AugmentationNode,
to fall into DataContainerNode case.

JIRA: YANGTOOLS-821
Change-Id: I5de7c814f0abe0aec31cced24ea09692bd63a584
Signed-off-by: Marek Gradzki <mgradzki@cisco.com>