Wrap lists and choices with StructuralContainerModificationStrategy
This achieves automatic lifecycle behavior, so that lists disappear
as soon as they have zero elements.

Tests are updated with the expected behavior -- most notably Bug4454,
which is losing enformement. This is because the min/max list is
inside a non-presence container, but root mandatory enforcement is
being ignored.

JIRA: YANGTOOLS-585
Change-Id: I68220043c0f335324e5a8788b2872f57b750aae9
Signed-off-by: Maros Marsalek <mmarsale@cisco.com>
Signed-off-by: Igor Foltin <ifoltin@cisco.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>