Wrap lists and choices with StructuralContainerModificationStrategy
This achieves automatic lifecycle behavior, so that lists disappear
as soon as they have zero elements.

Tests are updated with the expected behavior -- most notably Bug4454,
which is losing enformement. This is because the min/max list is
inside a non-presence container, but root mandatory enforcement is
being ignored.

FIXME: re-introduce a properly-initialized test (without mandatory
       masking).

JIRA: YANGTOOLS-575
Change-Id: I35bbb299c0d1b87a17855d230f609be18d0e2212
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>